combine cluster and schema server again
