+++ server/deploy/src/main/scala/cool/graph/deploy/server/ClusterServer.scala
@@ -13,9 +13,11 @@ import cool.graph.akkautil.http.Server
 import cool.graph.cuid.Cuid.createCuid
 import cool.graph.deploy.DeployMetrics
 import cool.graph.deploy.database.persistence.ProjectPersistence
-import cool.graph.deploy.schema.{DeployApiError, SchemaBuilder, SystemUserContext}
+import cool.graph.deploy.schema.{DeployApiError, InvalidProjectId, SchemaBuilder, SystemUserContext}
 import cool.graph.metrics.extensions.TimeResponseDirectiveImpl
+import cool.graph.shared.models.ProjectWithClientId
 import cool.graph.util.logging.{LogData, LogKey}
+import play.api.libs.json.Json
 import sangria.execution.{Executor, HandledException}
 import sangria.marshalling.ResultMarshaller
 import sangria.parser.QueryParser
@@ -28,6 +30,7 @@ import scala.util.{Failure, Success}
 
 case class ClusterServer(
     schemaBuilder: SchemaBuilder,
+    projectPersistence: ProjectPersistence,
     prefix: String = ""
 )(implicit system: ActorSystem, materializer: ActorMaterializer)
     extends Server
@@ -38,6 +41,7 @@ case class ClusterServer(
 
   val log: String => Unit = (msg: String) => logger.info(msg)
   val requestPrefix       = "cluster"
+  val server2serverSecret = sys.env.getOrElse("SCHEMA_MANAGER_SECRET", sys.error("SCHEMA_MANAGER_SECRET env var required but not found"))
 
   val innerRoutes = extractRequest { _ =>
     val requestId            = requestPrefix + ":cluster:" + createCuid()
@@ -107,10 +111,48 @@ case class ClusterServer(
           get {
             pathEnd {
               getFromResource("graphiql.html")
+            } ~ pathPrefix("schema") {
+              pathPrefix(Segment) { projectId =>
+                optionalHeaderValueByName("Authorization") {
+                  case Some(authorizationHeader) if authorizationHeader == s"Bearer $server2serverSecret" =>
+                    parameters('forceRefresh ? false) { forceRefresh =>
+                      complete(performSchemaRequest(projectId, forceRefresh, logRequestEnd))
+                    }
+
+                  case Some(h) =>
+                    println(s"Wrong Authorization Header supplied: '$h'")
+                    complete(Unauthorized -> "Wrong Authorization Header supplied")
+
+                  case None =>
+                    println("No Authorization Header supplied")
+                    complete(Unauthorized -> "No Authorization Header supplied")
+                }
+              }
             }
           }
       }
     }
+  }
+
+  def performSchemaRequest(projectId: String, forceRefresh: Boolean, requestEnd: (Option[String], Option[String]) => Unit) = {
+    getSchema(projectId, forceRefresh)
+      .map(res => OK -> res)
+      .andThen {
+        case _ => requestEnd(Some(projectId), None)
+      }
+      .recover {
+        case error: Throwable => BadRequest -> error.toString
+      }
+  }
+
+  def getSchema(projectId: String, forceRefresh: Boolean): Future[String] = {
+    import cool.graph.shared.models.ProjectJsonFormatter._
+    projectPersistence
+      .load(projectId)
+      .flatMap {
+        case None    => Future.failed(InvalidProjectId(projectId))
+        case Some(p) => Future.successful(Json.toJson(ProjectWithClientId(p, p.ownerId)).toString)
+      }
   }
 
   def healthCheck: Future[_] = Future.successful(())
