+++ server/single-server/src/main/resources/application.conf
@@ -49,5 +49,5 @@ client {
   connectionTimeout = 5000
 }
 
-schemaManagerEndpoint = ${SCHEMA_MANAGER_ENDPOINT}
+schemaManagerEndpoint = "http://172.16.123.1:9000/cluster/schema"
 schemaManagerSecret = ${SCHEMA_MANAGER_SECRET}
