+++ server/single-server/src/main/scala/cool/graph/singleserver/SingleServerMain.scala
@@ -5,7 +5,7 @@ import akka.stream.ActorMaterializer
 import cool.graph.akkautil.http.ServerExecutor
 import cool.graph.api.ApiDependenciesImpl
 import cool.graph.api.server.ApiServer
-import cool.graph.deploy.server.{ClusterServer, SchemaServer}
+import cool.graph.deploy.server.{ClusterServer}
 
 object SingleServerMain extends App {
   implicit val system          = ActorSystem("single-server")
@@ -19,9 +19,7 @@ object SingleServerMain extends App {
 
   ServerExecutor(
     port = port,
-    ClusterServer(singleServerDependencies.clusterSchemaBuilder, "cluster"),
-    SchemaServer(singleServerDependencies.projectPersistence, "cluster"),
-    ApiServer(singleServerDependencies.apiSchemaBuilder),
-    SchemaServer(singleServerDependencies.projectPersistence)
+    ClusterServer(singleServerDependencies.clusterSchemaBuilder, singleServerDependencies.projectPersistence, "cluster"),
+    ApiServer(singleServerDependencies.apiSchemaBuilder)
   ).startBlocking()
 }
